requirement.txt
blinker==1.6.3
click==8.1.7
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Uploads==0.2.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector==2.2.9
PyJWT==2.8.0
python-dotenv==1.0.0
Werkzeug==3.0.0
